Implement endpoint to get the bank statement for an account
This commit implements the endpoint that will fetch the bank statement
for a specific account. It allows the user by filtering it by 3, 15 or
30 days. It returns a json with the transactions in the the filter and a
past_balance with the total balance for transactions before the filter.

It is implemented by a class called BankStatement, that is responsible
to fetch this data in the Transactions table.