Implement a Warm Cache strategy
Imagine that we lost our cache data. This commit implements a behave in
Caching lib to allow we to warm cacheable resources. We can configure it
by setting a cache_options in the resource class and pass the parameters
to warm it. To execute the warm, this commit adds a Rake task `rake
cache:warm` that will execute this flow.